==> Data Analysis Process/02_DAP-Case2/00_assessing-case-2-solution.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd

fuel_08_df = pd.read_csv('all_alpha_08.csv')
fuel_18_df = pd.read_csv('all_alpha_18.csv')

print(fuel_08_df.shape)
fuel_08_df.head()

print(fuel_18_df.shape)
fuel_18_df.head()

fuel_08_df.isnull().sum()

fuel_18_df.isnull().sum()

fuel_08_df.duplicated().sum()

fuel_18_df.duplicated().sum()

fuel_08_df.dtypes

fuel_18_df.dtypes

fuel_08_df.SmartWay.nunique()

fuel_18_df.SmartWay.nunique()

fuel_08_df["Sales Area"].nunique()

fuel_18_df["Cert Region"].nunique()

fuel_08_df.Trans.nunique()

fuel_08_df.Trans.unique()

fuel_18_df.Trans.unique()

fuel_08_df.Fuel.unique()

fuel_18_df.Fuel.unique()
