==> Data Analysis Process/02_DAP-Case2/03_inspect_datatypes-case-2-solution.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08_v2.csv` and `data_18_v2.csv` to answer the quiz questions below regarding datatypes. You should've created these data files in the previous section: *Filter, Drop Nulls, Dedupe*.

import pandas as pd

fuel_08_df = pd.read_csv('data_08_v2.csv')
fuel_18_df = pd.read_csv('data_18_v2.csv')

fuel_08_df.dtypes

fuel_18_df.dtypes

fuel_08_df.cyl.unique()

fuel_18_df.cyl.unique()

fuel_08_df.air_pollution_score.unique()

fuel_18_df.air_pollution_score.unique()

Note that for the MPG columns, all of them are strings right now in both years.

fuel_08_df.greenhouse_gas_score.unique()

fuel_18_df.greenhouse_gas_score.unique()
